ai/requirement.txt
opencv-python
Flask 
opencv-python 
requests
paho-mqtt
